Update Configuration_adv.h
Changed XYZ Stepping to 256 from 16. Next will try to enable Hybrid Threshhold but I have had issues building when microsteps are at 256 and Hybrid is on. I believe this was from an old build envioronment and this branch is build from scratch on a fresh install of VS Code and Platform.io